restructured server to have view routers. added partners api

+++ server/src/views/trip_plan.py
@@ -1,20 +1,21 @@
 import json
-from typing import Optional, List, Union, Any
+from typing import Optional, List, Union, Any, Dict
 
-from fastapi import Depends, HTTPException
+from fastapi import Depends, HTTPException, APIRouter
 from pydantic import BaseModel
 
-from . import verify_access_key
-from .app import app
-from .db.database import get_session
-from .db.models import TripPlan, User
-from .utils.trip_plan.sunrise_and_sunset import flesh_out_itinerary
+from utils.auth import optional_user, verify_access_key
+from db.database import get_session
+from db.models import TripPlan, User
+from utils.trip_plan.sunrise_and_sunset import flesh_out_itinerary
 
 DT_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
 D_FORMAT = "%Y-%m-%d"
 
+router = APIRouter()
 
-@app.get("/api/v0/trip_plans")
+
+@router.get("/api/v0/trip_plans")
 async def get_trip_plans(user: User = Depends(verify_access_key)) -> list[TripPlan]:
     session = get_session()
     qs = session.query(TripPlan).filter(TripPlan.user_id == user.id)
@@ -41,29 +42,49 @@ def string_to_date(s):
     )
 
 
-@app.get("/api/v0/trip_plan/{trip_plan_id}")
+class GetTripPlanResponse(BaseModel):
+    id: str
+    user_id: str
+    name: str
+    packing_lists: Optional[Union[List, Dict]]
+    people: Union[List, Dict]
+    trails: Union[List, Dict]
+    dates: Dict
+    itinerary: Dict
+    editable: bool
+
+
+@router.get("/api/v0/trip_plan/{trip_plan_id}")
 async def get_trip_plan(
-    trip_plan_id: str, user: User = Depends(verify_access_key)
-) -> TripPlan:
+    trip_plan_id: str, user: User = Depends(optional_user)
+) -> GetTripPlanResponse:
     session = get_session()
-    qs = session.query(TripPlan).filter(
-        TripPlan.user_id == user.id,
-        TripPlan.id == trip_plan_id,
-    )
+    qs = session.query(TripPlan).filter(TripPlan.id == trip_plan_id)
     if qs.count() == 0:
-        raise HTTPException(status_code=404, detail="packing list not found")
+        raise HTTPException(status_code=404, detail="trip plan not found")
     trip_plan = qs.first()
     trip_plan.people = flesh_out_people(trip_plan.people)
+
+    editable = True
+    try:
+        validate_user_access(user, trip_plan)
+    except HTTPException:
+        editable = False
+
     trip_plan.itinerary = flesh_out_itinerary(trip_plan.trails, trip_plan.itinerary)
 
-    response = TripPlan(
+    response = GetTripPlanResponse(
+        user_id=trip_plan.user_id,
         name=trip_plan.name,
         id=trip_plan.id,
         dates=string_to_date(trip_plan.dates) if trip_plan.dates else None,
-        people=trip_plan.people,
-        trails=trip_plan.trails,
-        itinerary=trip_plan.itinerary,
+        people=trip_plan.people or [],
+        trails=trip_plan.trails or [],
+        itinerary=trip_plan.itinerary or [],
+        packing_lists=trip_plan.packing_lists or [],
+        editable=editable,
     )
+
     return response
 
 
@@ -101,17 +122,27 @@ def flesh_out_people(people):
     return fleshed_out
 
 
-@app.patch("/api/v0/trip_plan/{trip_plan_id}")
+def validate_user_access(user: User, trip_plan: TripPlan):
+    if user is None:
+        raise HTTPException(status_code=403, detail="access denied")
+    if trip_plan.user_id == user.id:
+        return True
+    for user in trip_plan.people:
+        if user["id"] == user.id:
+            return True
+    raise HTTPException(status_code=403, detail="access denied")
+
+
+@router.patch("/api/v0/trip_plan/{trip_plan_id}")
 async def update_trip_plan(
     trip_plan_id: str, request: TripPlanRequest, user: User = Depends(verify_access_key)
 ) -> TripPlan:
     session = get_session()
-    qs = session.query(TripPlan).filter(
-        TripPlan.user_id == user.id, TripPlan.id == trip_plan_id
-    )
+    qs = session.query(TripPlan).filter(TripPlan.id == trip_plan_id)
     if qs.count() == 0:
-        raise HTTPException(status_code=404, detail="packing list not found")
+        raise HTTPException(status_code=404, detail="trip plan not found")
     trip_plan = qs.first()
+    validate_user_access(user, trip_plan)
 
     # update values
     trip_plan.name = request.name
@@ -135,20 +166,20 @@ async def update_trip_plan(
     return response
 
 
-@app.delete("/api/v0/trip_plan/{trip_plan_id}")
+@router.delete("/api/v0/trip_plan/{trip_plan_id}")
 async def remove_trip_plan(trip_plan_id: str, user: User = Depends(verify_access_key)):
     session = get_session()
-    qs = session.query(TripPlan).filter(
-        TripPlan.user_id == user.id, TripPlan.id == trip_plan_id
-    )
+    qs = session.query(TripPlan).filter(TripPlan.id == trip_plan_id)
     if qs.count() == 0:
-        raise HTTPException(status_code=404, detail="packing list not found")
+        raise HTTPException(status_code=404, detail="trip plan not found")
     trip_plan = qs.first()
+    validate_user_access(user, trip_plan)
+
     session.delete(trip_plan)
     session.commit()
 
 
-@app.post("/api/v0/trip_plan")
+@router.post("/api/v0/trip_plan")
 async def create_trip_plan(user: User = Depends(verify_access_key)) -> str:
     trip_plan = TripPlan.new(name="new trip plan", user=user)
     session = get_session()
